feat(engine): generate terminate end event compatible execution steps part 3
This commit causes the BlockSequenceBuilder to stop generating execution segments once we have reached a terminate end event.

The BlockSequenceBuilder contains a sequence of blocks which all generate their own execution path segments. Imagine we have 2 blocks inside of the block sequence builder:

| task A -> task B |

If task A has a boundary event which has reached a terminate end event, task B should not execute as they are in the same flow scope. In order to prevent generating/appending the execution steps of task B we will break the loop if the segment of task A has reached a terminate end event.